refactor configs
